feat: move MCP Worker logic to /mcp-worker
