+++ mcp-worker/package.json
@@ -20,8 +20,9 @@
         "zod": "^3.24.2"
     },
     "devDependencies": {
-        "@types/node": "^24.0.10",
-        "typescript": "^5.8.3",
+        "@cloudflare/workers-types": "^4.20250726.0",
+        "@types/node": "^18.19.68",
+        "typescript": "^5.7.2",
         "wrangler": "^4.22.0"
     }
 }
