+++ mcp-worker/tsconfig.json
@@ -20,8 +20,7 @@
         "sourceMap": true
     },
     "include": [
-        "src/**/*",
-        "../src/mcp/worker/**/*"
+        "src/**/*"
     ],
     "exclude": [
         "node_modules",
