+++ mcp-worker/wrangler.toml
@@ -25,15 +25,15 @@ API_BASE_URL = "https://api.devcycle.com"
 AUTH0_DOMAIN = "auth.devcycle.com"
 AUTH0_AUDIENCE = "https://api.devcycle.com/"
 AUTH0_SCOPE = "openid profile email offline_access"
-# DEFAULT_PROJECT_KEY = ""  # Optional: Set a default project for single-project deployments
+
 
 # Secrets (set via: wrangler secret put <SECRET_NAME>)
 # AUTH0_CLIENT_ID
 # AUTH0_CLIENT_SECRET
 
-# AI binding for future features
-[ai]
-binding = "AI"
+# AI binding for future features (not currently used)
+# [ai]
+# binding = "AI"
 
 # Observability
 [observability]
