+++ tsconfig.json
@@ -24,5 +24,5 @@
         "tsBuildInfoFile": "./dist/.tsbuildinfo"
     },
     "include": ["src/**/*"],
-    "exclude": ["src/**/*.test.ts", "src/mcp/worker/**/*", "dist", "node_modules", "bin"]
+    "exclude": ["src/**/*.test.ts", "dist", "node_modules", "bin"]
 }
