+++ tsconfig.mcp.json
@@ -2,6 +2,7 @@
     "extends": "./tsconfig.json",
     "compilerOptions": {
         "outDir": "./dist/mcp-shared",
+        "rootDir": "./src",
         "declaration": true,
         "declarationMap": true,
         "sourceMap": true
@@ -9,7 +10,6 @@
     "include": [
         "src/mcp/tools/**/*",
         "src/mcp/api/interface.ts",
-        "src/mcp/worker/**/*",
         "src/api/**/*"
     ],
     "exclude": [
